+++ src/gitlab_watchman/models/note.py
@@ -1,6 +1,8 @@
 from dataclasses import dataclass
+from datetime import datetime
 
 from gitlab_watchman.models import user
+from gitlab_watchman.utils import convert_to_utc_datetime
 
 
 @dataclass(slots=True)
@@ -12,15 +14,15 @@ class Note(object):
     body: str
     attachment: str or bool
     author: user.User
-    created_at: str
-    updated_at: str
+    created_at: datetime | None
+    updated_at: datetime | None
     system: str
     noteable_id: str
     noteable_type: str
     commit_id: str
     resolvable: bool
     resolved_by: user.User
-    resolved_at: str
+    resolved_at: datetime | None
     confidential: str
     noteable_iid: str
     command_changes: str
@@ -39,21 +41,26 @@ def create_from_dict(note_dict: dict) -> Note:
     else:
         resolved_by = None
 
+    if note_dict.get('author'):
+        author = user.create_from_dict(note_dict.get('author', {}))
+    else:
+        author = None
+
     return Note(
         id=note_dict.get('id'),
         type=note_dict.get('type'),
         body=note_dict.get('body'),
         attachment=note_dict.get('attachment'),
-        author=user.create_from_dict(note_dict.get('author', {})),
-        created_at=note_dict.get('created_at'),
-        updated_at=note_dict.get('updated_at'),
+        author=author,
+        created_at=convert_to_utc_datetime(note_dict.get('created_at')),
+        updated_at=convert_to_utc_datetime(note_dict.get('updated_at')),
         system=note_dict.get('system'),
         noteable_id=note_dict.get('noteable_id'),
         noteable_type=note_dict.get('noteable_type'),
         commit_id=note_dict.get('commit_id'),
         resolvable=note_dict.get('resolvable'),
         resolved_by=resolved_by,
-        resolved_at=note_dict.get('resolved_at'),
+        resolved_at=convert_to_utc_datetime(note_dict.get('resolved_at')),
         confidential=note_dict.get('confidential'),
         noteable_iid=note_dict.get('noteable_iid'),
         command_changes=note_dict.get('command_changes'),