+++ src/gitlab_watchman/watchman_processor.py
@@ -3,14 +3,16 @@ import multiprocessing
 import re
 import time
 import traceback
+import hashlib
+from multiprocessing import Queue
 from dataclasses import dataclass
-from typing import List, Dict
+from typing import List, Dict, Optional
 
 from requests.exceptions import SSLError
 
 from gitlab_watchman.clients.gitlab_client import GitLabAPIClient
 from gitlab_watchman.exceptions import GitLabWatchmanGetObjectError, GitLabWatchmanAuthenticationError
-from gitlab_watchman.loggers import JSONLogger, StdoutLogger
+from gitlab_watchman.loggers import JSONLogger, StdoutLogger, init_logger
 from gitlab_watchman.models import (
     signature,
     note,
@@ -26,7 +28,11 @@ from gitlab_watchman.models import (
     project,
     group
 )
-from gitlab_watchman.utils import convert_time, deduplicate_results, split_to_chunks
+from gitlab_watchman.utils import (
+    convert_to_epoch,
+    deduplicate_results,
+    split_to_chunks
+)
 
 ALL_TIME = calendar.timegm(time.gmtime()) + 1576800000
 
@@ -34,14 +40,14 @@ ALL_TIME = calendar.timegm(time.gmtime()) + 1576800000
 @dataclass
 class WorkerArgs:
     """ Dataclass for multiprocessing arguments """
-
     gitlab_client: GitLabAPIClient
     search_result_list: List[Dict]
     regex: re.Pattern[str]
     timeframe: int
     results_list: List[Dict]
     verbose: bool
-    log_handler: JSONLogger | StdoutLogger
+    log_queue: Optional[Queue] = None
+    log_handler: Optional[JSONLogger | StdoutLogger] = None
 
 
 def initiate_gitlab_connection(token: str,
@@ -84,102 +90,103 @@ def find_group_owners(group_members: List[Dict]) -> List[Dict]:
     return member_list
 
 
-def find_user_owner(user_list: List[Dict]) -> List[Dict]:
-    """ Return user who owns a namespace
-
-    Args:
-        user_list: List of users
-    Returns:
-        List of formatted users owning a namespace
-    """
-
-    owner_list = []
-    for user in user_list:
-        owner_list.append({
-            'user_id': user.get('id'),
-            'name': user.get('name'),
-            'username': user.get('username'),
-            'state': user.get('state')
-        })
-
-    return owner_list
+def log_listener(log_queue: Queue, logging_type: str, debug: bool):
+    log_handler = init_logger(logging_type, debug)
+    while True:
+        record = log_queue.get()
+        if record is None:
+            break
+        level, message = record
+        log_handler.log(level, message)
 
 
 def search(gitlab: GitLabAPIClient,
-           log_handler: StdoutLogger | JSONLogger,
+           logging_type: str,
+           log_handler: JSONLogger | StdoutLogger,
+           debug: bool,
            sig: signature.Signature,
            scope: str,
            verbose: bool,
            timeframe: int = ALL_TIME) -> List[Dict] | None:
-    """ Uses the Search API to get search results for the given scope. These results are then split into (No of cores -
-    1) number of chunks, and Multiprocessing is then used to concurrently filter them against Regex using the relevant
-    worker function
-
-    Args:
-        gitlab: GitLab API object
-        log_handler: Logger object for outputting results
-        sig: Signature object
-        scope: What sort of GitLab objects to search
-        verbose: Whether to use verbose logging or not
-        timeframe: Timeframe to search in
-    Returns:
-        List of JSON formatted results if any are found
-    """
-
     results = []
+
+    if logging_type == 'json':
+        log_queue = Queue()
+        log_process = multiprocessing.Process(target=log_listener, args=(log_queue, logging_type, debug))
+        log_process.start()
 
     for query in sig.search_strings:
         for pattern in sig.patterns:
             regex = re.compile(pattern)
             search_results = gitlab.global_search(query, search_scope=scope)
             query_formatted = query.replace('"', '')
-            log_handler.log('INFO',
-                            f'{len(search_results)} {scope} found matching search term: {query_formatted}')
-            result = multiprocessing.Manager().list()
+            if search_results:
+                if logging_type == 'json':
+                    log_queue.put(('INFO', f'{len(search_results)} {scope} found matching search term: {query_formatted}'))
+                else:
+                    log_handler.log('INFO', f'{len(search_results)} {scope} found matching search term: {query_formatted}')
 
-            chunks = multiprocessing.cpu_count() - 1
-            list_of_chunks = split_to_chunks(search_results, chunks)
+                result = multiprocessing.Manager().list()
 
-            processes = []
+                chunks = multiprocessing.cpu_count() - 1
+                list_of_chunks = split_to_chunks(search_results, chunks)
 
-            target_func_dict = {
-                'blobs': _blob_worker,
-                'wiki_blobs': _wiki_blob_worker,
-                'commits': _commit_worker,
-                'snippet_titles': _snippet_worker,
-                'issues': _issue_worker,
-                'milestones': _milestone_worker,
-                'merge_requests': _merge_request_worker,
-                'notes': _note_worker,
-            }
-            target_func = target_func_dict.get(scope, _blob_worker)
+                processes = []
 
-            for search_list in list_of_chunks:
-                multipro_args = WorkerArgs(
-                    gitlab_client=gitlab,
-                    search_result_list=search_list,
-                    regex=regex,
-                    timeframe=timeframe,
-                    results_list=result,
-                    verbose=verbose,
-                    log_handler=log_handler
-                )
-                p = multiprocessing.Process(target=target_func,
-                                            args=(multipro_args,))
-                processes.append(p)
-                p.start()
+                target_func_dict = {
+                    'blobs': _blob_worker,
+                    'wiki_blobs': _wiki_blob_worker,
+                    'commits': _commit_worker,
+                    'snippet_titles': _snippet_worker,
+                    'issues': _issue_worker,
+                    'milestones': _milestone_worker,
+                    'merge_requests': _merge_request_worker,
+                    'notes': _note_worker,
+                }
+                target_func = target_func_dict.get(scope, _blob_worker)
 
-            for process in processes:
-                process.join()
+                for search_list in list_of_chunks:
+                    multipro_args = WorkerArgs(
+                        gitlab_client=gitlab,
+                        search_result_list=search_list,
+                        regex=regex,
+                        timeframe=timeframe,
+                        results_list=result,
+                        verbose=verbose
+                    )
+                    if logging_type == 'json':
+                        multipro_args.log_queue = log_queue
+                    else:
+                        multipro_args.log_handler = log_handler
+                    p = multiprocessing.Process(target=target_func, args=(multipro_args,))
+                    processes.append(p)
+                    p.start()
 
-            results.append(list(result))
+                for process in processes:
+                    process.join()
 
+                results.append(list(result))
+            else:
+                if logging_type == 'json':
+                    log_queue.put(('INFO', f'No {scope} found matching search term: {query_formatted}'))
+                else:
+                    log_handler.log('INFO', f'No {scope} found matching search term: {query_formatted}')
     if results:
         results = deduplicate_results([item for sublist in results for item in sublist])
-        log_handler.log('INFO', f'{len(results)} total matches found after filtering')
+        if logging_type == 'json':
+            log_queue.put(('INFO', f'{len(results)} total matches found after filtering'))
+            log_queue.put(None)
+            log_process.join()
+        else:
+            log_handler.log('INFO', f'{len(results)} total matches found after filtering')
         return results
     else:
-        log_handler.log('INFO', 'No matches found after filtering')
+        if logging_type == 'json':
+            log_queue.put(('INFO', 'No matches found after filtering'))
+            log_queue.put(None)
+            log_process.join()
+        else:
+            log_handler.log('INFO', 'No matches found after filtering')
 
 
 def _populate_project_owners(gitlab: GitLabAPIClient,
@@ -224,30 +231,36 @@ def _blob_worker(args: WorkerArgs) -> List[Dict]:
     """
 
     now = calendar.timegm(time.gmtime())
-    for b in args.search_result_list:
+    for blob_dict in args.search_result_list:
         try:
-            blob_object = blob.create_from_dict(b)
+            blob_object = blob.create_from_dict(blob_dict)
             project_object = project.create_from_dict(args.gitlab_client.get_project(blob_object.project_id))
             file_object = file.create_from_dict(
                 args.gitlab_client.get_file(blob_object.project_id, blob_object.path, blob_object.ref))
             if file_object:
                 commit_object = commit.create_from_dict(
                     args.gitlab_client.get_commit(blob_object.project_id, file_object.commit_id))
-                if convert_time(commit_object.committed_date) > (now - args.timeframe) and args.regex.search(
+                if convert_to_epoch(commit_object.committed_date) > (now - args.timeframe) and args.regex.search(
                         str(blob_object.data)):
                     match_string = args.regex.search(str(blob_object.data)).group(0)
                     if not args.verbose:
                         setattr(blob_object, 'data', None)
+                    watchman_id = hashlib.md5(f'{match_string}.{file_object.file_path}'.encode()).hexdigest()
                     args.results_list.append({
                         'match_string': match_string,
                         'blob': blob_object,
                         'commit': commit_object,
                         'project': _populate_project_owners(args.gitlab_client, project_object),
-                        'file': file_object
+                        'file': file_object,
+                        'watchman_id': watchman_id
                     })
         except GitLabWatchmanGetObjectError as e:
-            args.log_handler.log('WARNING', e)
-            args.log_handler.log('DEBUG', traceback.format_exc())
+            if args.log_handler:
+                args.log_handler.log('WARNING', e)
+                args.log_handler.log('DEBUG', traceback.format_exc())
+            else:
+                args.log_queue.put('WARNING', e)
+                args.log_queue.put('DEBUG', traceback.format_exc())
     return args.results_list
 
 
@@ -255,23 +268,23 @@ def _wiki_blob_worker(args: WorkerArgs) -> List[Dict]:
     """ MULTIPROCESSING WORKER - Iterates through a list of wiki_blobs to find matches against the regex.
 
     Args:
-        args (WorkerArgs): Multiprocessing arguments containing the
+        args: Multiprocessing arguments containing the
                                     GitLab client, search list, regex pattern,
                                     timeframe, results list, verbosity flag, and log handler.
     Returns:
         List[Dict]: Multiprocessing list to be combined by the parent process.
     """
 
-    for wb in args.search_result_list:
+    for wb_dict in args.search_result_list:
         try:
-            wikiblob_object = wiki_blob.create_from_dict(wb)
+            wikiblob_object = wiki_blob.create_from_dict(wb_dict)
             project_wiki = False
             group_wiki = False
-            if wb.get('project_id'):
-                project_object = project.create_from_dict(args.gitlab_client.get_project(wb.get('project_id')))
+            if wb_dict.get('project_id'):
+                project_object = project.create_from_dict(args.gitlab_client.get_project(wb_dict.get('project_id')))
                 project_wiki = True
-            if wb.get('group_id'):
-                group_object = group.create_from_dict(args.gitlab_client.get_group(wb.get('group_id')))
+            if wb_dict.get('group_id'):
+                group_object = group.create_from_dict(args.gitlab_client.get_group(wb_dict.get('group_id')))
                 group_wiki = True
 
             if args.regex.search(
@@ -279,11 +292,13 @@ def _wiki_blob_worker(args: WorkerArgs) -> List[Dict]:
                 match_string = args.regex.search(str(wikiblob_object.data)).group(0)
                 if not args.verbose:
                     setattr(wikiblob_object, 'data', None)
+                watchman_id = hashlib.md5(f'{match_string}.{wikiblob_object.path}'.encode()).hexdigest()
                 results_dict = {
                     'match_string': match_string,
                     'wiki_blob': wikiblob_object,
                     'group_wiki': group_wiki,
                     'project_wiki': project_wiki,
+                    'watchman_id': watchman_id
                 }
                 if project_wiki:
                     results_dict['project'] = _populate_project_owners(args.gitlab_client, project_object)
@@ -291,8 +306,12 @@ def _wiki_blob_worker(args: WorkerArgs) -> List[Dict]:
                     results_dict['group'] = group_object
                 args.results_list.append(results_dict)
         except GitLabWatchmanGetObjectError as e:
-            args.log_handler.log('WARNING', e)
-            args.log_handler.log('DEBUG', traceback.format_exc())
+            if args.log_handler:
+                args.log_handler.log('WARNING', e)
+                args.log_handler.log('DEBUG', traceback.format_exc())
+            else:
+                args.log_queue.put('WARNING', e)
+                args.log_queue.put('DEBUG', traceback.format_exc())
     return args.results_list
 
 
@@ -309,20 +328,27 @@ def _commit_worker(args: WorkerArgs) -> List[Dict]:
 
     now = calendar.timegm(time.gmtime())
 
-    for c in args.search_result_list:
+    for commit_dict in args.search_result_list:
         try:
-            commit_object = commit.create_from_dict(c)
-            if convert_time(commit_object.committed_date) > (now - args.timeframe) and \
+            commit_object = commit.create_from_dict(commit_dict)
+            if convert_to_epoch(commit_object.committed_date) > (now - args.timeframe) and \
                     args.regex.search(str(commit_object.message)):
                 project_object = project.create_from_dict(args.gitlab_client.get_project(commit_object.project_id))
+                match_string = args.regex.search(str(commit_object.message)).group(0)
+                watchman_id = hashlib.md5(f'{match_string}.{commit_object.id}'.encode()).hexdigest()
                 args.results_list.append({
-                    'match_string': args.regex.search(str(commit_object.message)).group(0),
+                    'match_string': match_string,
                     'commit': commit_object,
-                    'project': _populate_project_owners(args.gitlab_client, project_object)
+                    'project': _populate_project_owners(args.gitlab_client, project_object),
+                    'watchman_id': watchman_id
                 })
         except GitLabWatchmanGetObjectError as e:
-            args.log_handler.log('WARNING', e)
-            args.log_handler.log('DEBUG', traceback.format_exc())
+            if args.log_handler:
+                args.log_handler.log('WARNING', e)
+                args.log_handler.log('DEBUG', traceback.format_exc())
+            else:
+                args.log_queue.put('WARNING', e)
+                args.log_queue.put('DEBUG', traceback.format_exc())
     return args.results_list
 
 
@@ -338,23 +364,29 @@ def _issue_worker(args: WorkerArgs) -> List[Dict]:
     """
 
     now = calendar.timegm(time.gmtime())
-    for i in args.search_result_list:
+    for issue_dict in args.search_result_list:
         try:
-            issue_object = issue.create_from_dict(i)
-            if convert_time(issue_object.updated_at) > (now - args.timeframe) and \
+            issue_object = issue.create_from_dict(issue_dict)
+            if convert_to_epoch(issue_object.updated_at) > (now - args.timeframe) and \
                     args.regex.search(str(issue_object.description)):
                 match_string = args.regex.search(str(issue_object.description)).group(0)
                 if not args.verbose:
                     setattr(issue_object, 'description', None)
                 project_object = project.create_from_dict(args.gitlab_client.get_project(issue_object.project_id))
+                watchman_id = hashlib.md5(f'{match_string}.{issue_object.id}'.encode()).hexdigest()
                 args.results_list.append({
                     'match_string': match_string,
                     'issue': issue_object,
-                    'project': _populate_project_owners(args.gitlab_client, project_object)
+                    'project': _populate_project_owners(args.gitlab_client, project_object),
+                    'watchman_id': watchman_id
                 })
         except GitLabWatchmanGetObjectError as e:
-            args.log_handler.log('WARNING', e)
-            args.log_handler.log('DEBUG', traceback.format_exc())
+            if args.log_handler:
+                args.log_handler.log('WARNING', e)
+                args.log_handler.log('DEBUG', traceback.format_exc())
+            else:
+                args.log_queue.put('WARNING', e)
+                args.log_queue.put('DEBUG', traceback.format_exc())
     return args.results_list
 
 
@@ -368,23 +400,29 @@ def _milestone_worker(args: WorkerArgs) -> List[Dict]:
     """
 
     now = calendar.timegm(time.gmtime())
-    for m in args.search_result_list:
+    for milestone_dict in args.search_result_list:
         try:
-            milestone_object = milestone.create_from_dict(m)
-            if convert_time(milestone_object.updated_at) > (now - args.timeframe) and \
+            milestone_object = milestone.create_from_dict(milestone_dict)
+            if convert_to_epoch(milestone_object.updated_at) > (now - args.timeframe) and \
                     args.regex.search(str(milestone_object.description)):
                 project_object = project.create_from_dict(args.gitlab_client.get_project(milestone_object.project_id))
                 match_string = args.regex.search(str(milestone_object.description)).group(0)
                 if not args.verbose:
                     setattr(milestone_object, 'description', None)
+                watchman_id = hashlib.md5(f'{match_string}.{milestone_object.id}'.encode()).hexdigest()
                 args.results_list.append({
                     'match_string': match_string,
                     'milestone': milestone_object,
-                    'project': _populate_project_owners(args.gitlab_client, project_object)
+                    'project': _populate_project_owners(args.gitlab_client, project_object),
+                    'watchman_id': watchman_id
                 })
         except GitLabWatchmanGetObjectError as e:
-            args.log_handler.log('WARNING', e)
-            args.log_handler.log('DEBUG', traceback.format_exc())
+            if args.log_handler:
+                args.log_handler.log('WARNING', e)
+                args.log_handler.log('DEBUG', traceback.format_exc())
+            else:
+                args.log_queue.put('WARNING', e)
+                args.log_queue.put('DEBUG', traceback.format_exc())
     return args.results_list
 
 
@@ -400,23 +438,29 @@ def _merge_request_worker(args: WorkerArgs) -> List[Dict]:
     """
 
     now = calendar.timegm(time.gmtime())
-    for mr in args.search_result_list:
+    for mr_dict in args.search_result_list:
         try:
-            mr_object = merge_request.create_from_dict(mr)
-            if convert_time(mr_object.updated_at) > (now - args.timeframe) and \
+            mr_object = merge_request.create_from_dict(mr_dict)
+            if convert_to_epoch(mr_object.updated_at) > (now - args.timeframe) and \
                     args.regex.search(str(mr_object.description)):
                 project_object = project.create_from_dict(args.gitlab_client.get_project(mr_object.project_id))
                 match_string = args.regex.search(str(mr_object.description)).group(0)
                 if not args.verbose:
                     setattr(mr_object, 'description', None)
+                watchman_id = hashlib.md5(f'{match_string}.{mr_object.id}'.encode()).hexdigest()
                 args.results_list.append({
                     'match_string': match_string,
                     'merge_request': mr_object,
-                    'project': _populate_project_owners(args.gitlab_client, project_object)
+                    'project': _populate_project_owners(args.gitlab_client, project_object),
+                    'watchman_id': watchman_id
                 })
         except GitLabWatchmanGetObjectError as e:
-            args.log_handler.log('WARNING', e)
-            args.log_handler.log('DEBUG', traceback.format_exc())
+            if args.log_handler:
+                args.log_handler.log('WARNING', e)
+                args.log_handler.log('DEBUG', traceback.format_exc())
+            else:
+                args.log_queue.put('WARNING', e)
+                args.log_queue.put('DEBUG', traceback.format_exc())
     return args.results_list
 
 
@@ -433,18 +477,24 @@ def _note_worker(args: WorkerArgs) -> List[Dict]:
 
     now = calendar.timegm(time.gmtime())
     try:
-        for n in args.search_result_list:
-            note_object = note.create_from_dict(n)
-            if convert_time(note_object.created_at) > (now - args.timeframe) and \
+        for note_dict in args.search_result_list:
+            note_object = note.create_from_dict(note_dict)
+            if convert_to_epoch(note_object.created_at) > (now - args.timeframe) and \
                     args.regex.search(str(note_object.body)):
                 match_string = args.regex.search(str(note_object.body)).group(0)
+                watchman_id = hashlib.md5(f'{match_string}.{note_object.id}'.encode()).hexdigest()
                 args.results_list.append({
                     'note': note_object,
-                    'match_string': match_string
+                    'match_string': match_string,
+                    'watchman_id': watchman_id
                 })
     except GitLabWatchmanGetObjectError as e:
-        args.log_handler.log('WARNING', e)
-        args.log_handler.log('DEBUG', traceback.format_exc())
+        if args.log_handler:
+            args.log_handler.log('WARNING', e)
+            args.log_handler.log('DEBUG', traceback.format_exc())
+        else:
+            args.log_queue.put('WARNING', e)
+            args.log_queue.put('DEBUG', traceback.format_exc())
     return args.results_list
 
 
@@ -463,20 +513,25 @@ def _snippet_worker(args: WorkerArgs) -> List[Dict]:
     for snippet_dict in args.search_result_list:
         try:
             snippet_object = snippet.create_from_dict(snippet_dict)
-            if convert_time(snippet_object.created_at) > (now - args.timeframe) and \
+            if convert_to_epoch(snippet_object.created_at) > (now - args.timeframe) and \
                     (args.regex.search(str(snippet_object.title)) or args.regex.search(str(snippet_object.description))):
                 if args.regex.search(str(snippet_object.title)):
                     match_string = args.regex.search(str(snippet_object.title)).group(0)
                 else:
                     match_string = args.regex.search(str(snippet_object.description)).group(0)
-
                 if not args.verbose:
                     setattr(snippet_object, 'description', None)
+                watchman_id = hashlib.md5(f'{match_string}.{snippet_object.id}'.encode()).hexdigest()
                 args.results_list.append({
                     'snippet': snippet_object,
-                    'match_string': match_string
+                    'match_string': match_string,
+                    'watchman_id': watchman_id
                 })
         except GitLabWatchmanGetObjectError as e:
-            args.log_handler.log('WARNING', e)
-            args.log_handler.log('DEBUG', traceback.format_exc())
+            if args.log_handler:
+                args.log_handler.log('WARNING', e)
+                args.log_handler.log('DEBUG', traceback.format_exc())
+            else:
+                args.log_queue.put('WARNING', e)
+                args.log_queue.put('DEBUG', traceback.format_exc())
     return args.results_list